requirements.txt
anthropic>=0.34.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
jupyter>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
